Rename brain_tree to aabb_tree
